Move replacer to utils
